% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{bayesianbinaryprobit}
\alias{bayesianbinaryprobit}
\title{Binary Probit Regression}
\usage{
bayesianbinaryprobit(formula, data, covar_prior, num_mcmc = 10000,
  burn_in = 1000, thinning = 1)
}
\arguments{
\item{formula}{: formula for the model}

\item{data}{: data frame containing input data}

\item{covar_prior:}{covariance matrix of 0 mean gaussian priors for coefficients}

\item{num_mcmc:}{Number of MCMC samples}

\item{burn_in:}{Number of burn in samples}

\item{thinning:}{Thinning interval}
}
\value{
coeffs: data frame containing coeffcient estimates and their standard deviation
}
\description{
Binary Probit Regression function
}
