% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{bayesprobitmcmc}
\alias{bayesprobitmcmc}
\title{MCMC chain for Bayesian Binary Probit Regression}
\usage{
bayesprobitmcmc(y, X, covar_prior, num_mcmc = 10000, burn_in = 1000,
  thinning = 1)
}
\arguments{
\item{y}{: vector with either 0 or 1}

\item{X}{: matrix with independent variables as columns, and first column is a vector of 1s}

\item{covar_prior:}{covariance matrix of 0 mean gaussian priors for coefficients}

\item{num_mcmc:}{Number of MCMC samples}

\item{burn_in:}{Number of burn in samples}

\item{thinning:}{Thinning interval}
}
\value{
coeffs: data frame containing coeffcient estimates and their standard deviation
}
\description{
Run MCMC chain and get estimates for Logistic Regression Coefficients
}
