% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{choleskylower}
\alias{choleskylower}
\title{Cholesky Decomposition}
\usage{
choleskylower(mat)
}
\arguments{
\item{mat:}{a square matrix}
}
\value{
L: cholesky lower traingular decomposition of mat
}
\description{
Lower Triangular Cholesky decomposition of a matrix
}
