% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{gausseliminationinverse}
\alias{gausseliminationinverse}
\title{Matrix Inversion}
\usage{
gausseliminationinverse(mat)
}
\arguments{
\item{mat:}{a square matrix}
}
\value{
matinv: inverse of mat
}
\description{
Invert a matrix by Gauss Elimination method
}
