% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{gaussiansamplesbyCLT}
\alias{gaussiansamplesbyCLT}
\title{Samples From Gaussian Distribution}
\usage{
gaussiansamplesbyCLT(num_samples, mean_norm = 0, sd_norm = 1,
  num_uniform_samples = 50)
}
\arguments{
\item{mean_norm}{: Mean of Gaussian distribution}

\item{sd_norm}{: Standard Deviation of Gaussian distribution}

\item{num_samples:}{Number of samples to be generated}

\item{num_uniform_samples:}{Number of Uniform samples to be used for generating Gaussian sample by averaging}
}
\value{
samples: Gaussian samples
}
\description{
Generate samples from Gaussian distribution
Uses central limit theorem to generate gaussian samples from uniform samples
}
