% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianProbitRegression.R
\name{samplegaussianonesided}
\alias{samplegaussianonesided}
\title{Samples From One Sided Gaussian Distribution}
\usage{
samplegaussianonesided(num_samples, mean_norm = 0, sd_norm = 1,
  cut_off = 0, side = c("right", "left"), num_uniform_samples = 50)
}
\arguments{
\item{mean_norm}{: Mean of Gaussian distribution}

\item{sd_norm}{: Standard Deviation of Gaussian distribution}

\item{cut_off}{: Cut off value for truncation}

\item{side}{: right or left based on which sideddistribution to be generated}

\item{num_samples:}{Number of samples to be generated}

\item{num_uniform_samples:}{Number of Uniform samples to be used for generating Gaussian sample by averaging}
}
\value{
samples: one sided Gaussian samples
}
\description{
Generate samples from one sided Gaussian distribution
Uses central limit theorem to generate gaussian samples from uniform samples
}
